refactor: 清理重复代码，修复约定违反与 Wayland 尺寸缺陷 (v4.3.3)
- Pointer::click 默认实现 / DirState / key_pos 泛型 / GridFilter 合并，去重
- warp 坐标 i16 -> i32（多屏坐标不再截断）
- WlrBackend 枚举真实 wl_output，Mode 只取 CURRENT
- 修复 windows MSG zeroed（违反 AGENTS.md）、device_perm uid 解析

+++ src/overlay.rs
@@ -54,7 +54,7 @@ pub trait OverlayBackend {
     fn upload(&self, idx: usize, skia: &SkiaPixmap) -> Result<()>;
     fn show_all(&self) -> Result<()>;
     fn redraw_all(&self) -> Result<()>;
-    fn pointer_warp(&self, x: i16, y: i16) -> Result<()>;
+    fn pointer_warp(&self, x: i32, y: i32) -> Result<()>;
 }
 
 /// The active overlay backend for the current platform.
